Refactor distributed.rs into coordinator module and explicitly enable task metrics collection
