+++ src/coordinator/mod.rs
@@ -1,0 +1,7 @@
+mod distributed;
+mod metrics_store;
+mod prepare_static_plan;
+mod task_spawner;
+
+pub use distributed::DistributedExec;
+pub(crate) use metrics_store::MetricsStore;
